separate-product-inventory
